Using HTML of postbody instead of text now
The original poster variable now uses contains because it could have
multiple classes, which meant it wasnt finding anything when there was a
moderator who had started a thread.

It now extracts the HTML of the postbody td, not the text. Appropriate
changes were made to the xslt sheet also to ensure that the HTML would
actually render. It does not work in Firefox, Chrome - But it does work
in Internet Explorer, which is what I was using to display these sites
anyway. Some cleaning is done to this HTML too, but I dont think it
works properly.

Output body is put inside a CDATA tag, because that's apparently
necessary.