Established update functionality.
Small changes:
- Created a clean_post function.
- Created a create_post_xml function.
- Renamed loop variable j to p, since it's looping over pages.

The strategy established in previous commit was followed pretty
well. However the changes made follows,

Added a continue_from parameter to get_post_information. This
continue_from parameter is used to cut the list of pages, as
an update function is not very helpful if you just iterate
over the entire thread again. The continue_from parameter is in
reality the last_post added to the xml, so we look at it's page number.
This page number's index is looked up in the list of pages, and then we
slice the list using that. After that, the get_post_information function
continues as usual.

Get get_post_information function is called, as usual, from
get_thread after having checked if the update parameter is set
to True, and if it is we open the old xml file, look up the last
post made, and send that along as a variable. If the update
parameter is False, we just call it without the continue_from
parameter being set to anything.

Next change made is in create_xml, we added an update parameter
here aswell, because this needs to know if it's writing a whole
new thread, or if it's updating a previously established one.

So if the update parameter is set to True in create_xml, we open
old thread and now this is important: We create an XMLParser instance
first, this XMLParser instance has the remove_blank_text parameter set
this parser is then used to open the thread, do not we do not use
lxml.etree.fromstring here, we use lxml.etree.parse for a couple of
reasons.
The first being we can reuse the parse instance again to write,
which we would have to make an ElementTree object with the
fromstring etree object as a parameter. The second being that we
can pass the parser argument in, I don't know if you can also do
this with fromstring, but compounded with reason one this is preferable.
We set the remove_blank_text parameter in the parser for a reason,
if it's not set the parser retains the whitespace, and the pretty_printer
breaks if the whitespace is retained.

Having opened the thread, we find the last_post, filter out all the posts
we don't want based on it's post_number, e.g we want it to have a larger
post_number than the last_post. Then we find the posts element using
xpath and append our posts, which were transformed in the
list comprehension/filter.
Then we save the thread.

If the update parameter is set to False, we do it the old way.

NOTE: I'm rapidly losing overview of this mess. Next updates will
start making the entire thing OOP.